@@ -6,23 +6,23 @@
     "serve": "vite preview"
   },
   "dependencies": {
-    "leva": "^0.9.25",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "leva": "^0.9.35",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
-    "@vitejs/plugin-react": "^1.3.1",
-    "eslint": "^8.14.0",
+    "@types/react": "^18.2.79",
+    "@types/react-dom": "^18.2.25",
+    "@typescript-eslint/eslint-plugin": "<7",
+    "@typescript-eslint/parser": "<7",
+    "@vitejs/plugin-react": "^4.2.1",
+    "eslint": "<9",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.5.0",
-    "prettier": "^2.6.2",
-    "typescript": "^4.6.3",
-    "vite": "^2.9.6"
+    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^3.2.5",
+    "typescript": "^5.4.5",
+    "vite": "^5.2.10"
   },
   "prettier": {
     "trailingComma": "all"
