@@ -1,11 +1,11 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
+import { createRoot } from "react-dom/client";
 
-ReactDOM.render(
+const reactRoot = createRoot(document.getElementById("root")!);
+reactRoot.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById("root"),
 );
